solver: partition requirement parsing into its own file
